pyproject.toml
[tool.poetry]
name = "discord-bot-rag"
version = "2.0.0"
description = "Discord chatbot with conversational memory and RAG capabilities"
authors = ["Gabriel Ramos <gabriel@example.com>"]
readme = "README.md"
packages = [{include = "discord_bot", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.13"
discord-py = "^2.3.2"
aiohttp = "^3.9.0"
chromadb = "^0.5.0"
python-docx = "^1.2.0"
docx2txt = "^0.9"
openai = "^2.6.0"
pymupdf = "^1.26.5"
django = "^4.2.0"
django-cors-headers = "^4.0.0"
whitenoise = "^6.5.0"
psutil = "^5.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
ruff = "^0.0.285"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"